perf: fix unbounded parallel file reading in search index generation
Replace Promise.all(files.map(...)) with a sequential for...of loop to
prevent resource exhaustion (memory/file descriptors) when processing
a large number of documentation files.

Also standardizes indentation to 2 spaces within the loop body.
