+++ .github/workflows/deploy-platform.yml
@@ -1,8 +1,5 @@
 name: Deploy — GoldShore Platform
 
-permissions:
-  contents: read
-
 # Triggered on push to main OR manually from GitHub Actions tab
 on:
   push:
@@ -13,9 +10,6 @@ on:
         description: 'Skip deleting stub workers (goldshore-ai, gs-dynamic-worker)'
         required: false
         default: 'false'
-
-permissions:
-  contents: read
 
 jobs:
   # ─────────────────────────────────────────────
@@ -28,6 +22,8 @@ jobs:
       - uses: actions/checkout@v4
 
       - uses: pnpm/action-setup@v4
+        with:
+          version: 9
 
       - uses: actions/setup-node@v4
         with:
@@ -56,6 +52,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: pnpm/action-setup@v4
+        with: { version: 9 }
       - uses: actions/setup-node@v4
         with: { node-version: 22, cache: pnpm }
       - run: pnpm install --frozen-lockfile
@@ -76,6 +73,7 @@ jobs:
     steps:
       - uses: actions/checkout@v4
       - uses: pnpm/action-setup@v4
+        with: { version: 9 }
       - uses: actions/setup-node@v4
         with: { node-version: 22, cache: pnpm }
       - run: pnpm install --frozen-lockfile
