+++ apps/gs-web/src/utils/security.ts
@@ -33,33 +33,3 @@ export function sanitizeInput(str: string): string {
   if (typeof str !== "string") return "";
   return escapeHtml(str.trim());
 }
-
-/**
- * Validates that a request originated from the same origin as the application.
- * Checks Origin, Referer, and Sec-Fetch-Site headers.
- * @param request The incoming Request object.
- * @returns True if the request is from the same origin, false otherwise.
- */
-export const isSameOriginRequest = (request: Request) => {
-  const expectedOrigin = new URL(request.url).origin;
-  const originHeader = request.headers.get('origin');
-  if (originHeader) {
-    return originHeader === expectedOrigin;
-  }
-
-  const refererHeader = request.headers.get('referer');
-  if (refererHeader) {
-    try {
-      return new URL(refererHeader).origin === expectedOrigin;
-    } catch {
-      return false;
-    }
-  }
-
-  const fetchSite = request.headers.get('sec-fetch-site');
-  if (fetchSite) {
-    return fetchSite === 'same-origin' || fetchSite === 'none';
-  }
-
-  return false;
-};
