+++ apps/gs-web/tests/unit/lead-submissions.test.ts
@@ -1,8 +1,9 @@
 import { test } from 'node:test';
 import * as assert from 'node:assert/strict';
-import { isSameOriginRequest } from '../../src/utils/security.ts';
 
-test('isSameOriginRequest: accepts same-origin POSTs via Origin header', () => {
+import { __testing } from '../../src/pages/api/admin/lead-submissions.ts';
+
+test('accepts same-origin POSTs via Origin header', () => {
   const request = new Request('https://admin.goldshore.ai/api/admin/lead-submissions', {
     method: 'POST',
     headers: {
@@ -10,10 +11,10 @@ test('isSameOriginRequest: accepts same-origin POSTs via Origin header', () => {
     },
   });
 
-  assert.equal(isSameOriginRequest(request), true);
+  assert.equal(__testing.isSameOriginRequest(request), true);
 });
 
-test('isSameOriginRequest: rejects cross-site POSTs via Origin header', () => {
+test('rejects cross-site POSTs via Origin header', () => {
   const request = new Request('https://admin.goldshore.ai/api/admin/lead-submissions', {
     method: 'POST',
     headers: {
@@ -21,10 +22,10 @@ test('isSameOriginRequest: rejects cross-site POSTs via Origin header', () => {
     },
   });
 
-  assert.equal(isSameOriginRequest(request), false);
+  assert.equal(__testing.isSameOriginRequest(request), false);
 });
 
-test('isSameOriginRequest: falls back to Referer when Origin is absent', () => {
+test('falls back to Referer when Origin is absent', () => {
   const request = new Request('https://admin.goldshore.ai/api/admin/lead-submissions', {
     method: 'POST',
     headers: {
@@ -32,13 +33,13 @@ test('isSameOriginRequest: falls back to Referer when Origin is absent', () => {
     },
   });
 
-  assert.equal(isSameOriginRequest(request), true);
+  assert.equal(__testing.isSameOriginRequest(request), true);
 });
 
-test('isSameOriginRequest: rejects requests without browser same-origin context', () => {
+test('rejects requests without browser same-origin context', () => {
   const request = new Request('https://admin.goldshore.ai/api/admin/lead-submissions', {
     method: 'POST',
   });
 
-  assert.equal(isSameOriginRequest(request), false);
+  assert.equal(__testing.isSameOriginRequest(request), false);
 });
